LaundrySymbols moving styled text composables to another class
