+++ app/src/main/java/com/tomerpacific/laundry/StyledText.kt
@@ -1,0 +1,28 @@
+package com.tomerpacific.laundry
+
+import android.view.Gravity
+import android.widget.TextView
+import androidx.annotation.StringRes
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.ReadOnlyComposable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.viewinterop.AndroidView
+
+@Composable
+@ReadOnlyComposable
+fun textResource(@StringRes id: Int): CharSequence =
+    LocalContext.current.resources.getText(id)
+
+@Composable
+fun StyledText(text: CharSequence, modifier: Modifier = Modifier) {
+    AndroidView(
+        modifier = modifier,
+        factory = { context -> TextView(context) },
+        update = {
+            it.text = text
+            it.textSize = 20f
+            it.gravity = Gravity.CENTER
+        }
+    )
+}
