+++ app/src/main/java/com/tomerpacific/laundry/fragment/HowToDoLaundryFragment.kt
@@ -4,8 +4,6 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
-import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -30,21 +28,20 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.rememberDrawerState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.ComposeView
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.viewinterop.AndroidView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import com.tomerpacific.laundry.StyledText
 import com.tomerpacific.laundry.model.HowToDoLaundryCategories
 import com.tomerpacific.laundry.model.HowToDoLaundryCategory
+import com.tomerpacific.laundry.textResource
 import com.tomerpacific.laundry.viewmodel.MainViewModel
 import kotlinx.coroutines.launch
 
@@ -159,23 +156,5 @@ class HowToDoLaundryFragment: Fragment() {
                 }
             }
         }
-    }
-
-    @Composable
-    @ReadOnlyComposable
-    fun textResource(@StringRes id: Int): CharSequence =
-        LocalContext.current.resources.getText(id)
-
-    @Composable
-    fun StyledText(text: CharSequence, modifier: Modifier = Modifier) {
-        AndroidView(
-            modifier = modifier,
-            factory = { context -> TextView(context) },
-            update = {
-                it.text = text
-                it.textSize = 20f
-                it.gravity = android.view.Gravity.CENTER
-            }
-        )
     }
 }
